+++ NEWS.md
@@ -1,3 +1,11 @@
+# protocolhelper 0.4.13
+
+## Minor changes
+
+* internal function `render_release()` now also writes a `.zenodo.json` file
+* protocol authors are now added as contributors instead of creators in 
+  `.zenodo.json` file
+
 # protocolhelper 0.4.12
 
 ## Minor changes